feat(archive): реализовать функционал архивации записей
Добавлена возможность архивировать записи, чтобы скрыть их из основного списка без удаления. Включает изменения в логике и пользовательском интерфейсе.

Логика:
- В модель данных добавлен флаг IsArchive для отметки архивных записей.
- Реализован механизм перемещения записей в архив, который запрещает их последующее удаление.

UI:
- Добавлен новый popup для просмотра и управления архивированными записями.
- Кнопка вызова popup становится доступной только при наличии записей в архиве.
- Реализовано кастомное позиционирование и динамическое изменение высоты popup (расширяется вверх до `Page.ActualHeight - 150px`).
- Обновлены стили для ToggleButton в боковом меню для соответствия общему дизайну.

+++ apps/apis/EnigmaVault.SecretService.Api/EnigmaVault.SecretService.Application/Features/Secrets/Delete/DeleteSecretCommandHandler.cs
@@ -13,6 +13,14 @@ namespace EnigmaVault.SecretService.Application.Features.Secrets.Delete
 
         public async Task<Result> Handle(DeleteSecretCommand request, CancellationToken cancellationToken)
         {
+            var domain = await _secretRepository.GetByIdAsync(request.IdSecret, cancellationToken);
+
+            if (domain == null)
+                return Result.Failure(new Error(ErrorCode.NotFound, "Не удалось найти запись"));
+
+            if (domain.IsArchive)
+                return Result.Failure(new Error(ErrorCode.DeleteError, "Нельзя удалить запись которая находиться в архиве"));
+
             var isDelete = await _secretRepository.DeleteAsync(request.IdSecret);
 
             if (isDelete)
